+++ bundler/spec/support/path.rb
@@ -3,8 +3,12 @@
 require "pathname"
 require "rbconfig"
 
+require_relative "env"
+
 module Spec
   module Path
+    include Spec::Env
+
     def source_root
       @source_root ||= Pathname.new(ruby_core? ? "../../.." : "../..").expand_path(__dir__)
     end
@@ -255,10 +259,6 @@ module Spec
       contents = File.read(gemspec_file)
       contents.sub!(/(^\s+s\.required_ruby_version\s*=\s*)"[^"]+"/, %(\\1"#{version}"))
       File.open(gemspec_file, "w") {|f| f << contents }
-    end
-
-    def ruby_core?
-      !ENV["GEM_COMMAND"].nil?
     end
 
     def git_root
